requirements.txt
Django==1.8.7
django-flatblocks
Pillow
django-autoslug
